# contracts/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Data Structure Agent API
  version: 1.0.0
  description: Spring backend contract for the data structure learning agent.
servers:
  - url: /api/v1
    description: Relative base; usable from structify.cn and admin.structify.cn. Admin routes still require a database-backed ADMIN role.
security: []
...
paths:
  /auth/request-code:
    post:
      tags: [Auth]
      summary: Request an email verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestCodeRequest'
      responses:
        '200':
          description: Generic delivery response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationCodeDelivery'
        '400': {$ref: '#/components/responses/BadRequest'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a verified account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Authenticated account; also sets the HttpOnly session cookie
          headers:
            Set-Cookie:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated account; also sets the HttpOnly session cookie
          headers:
            Set-Cookie:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset a password with a reset-purpose verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password updated and a new session issued
          headers:
            Set-Cookie:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /admin/capabilities:
    get:
      tags: [Admin]
      summary: Return the current administrator's available management modules
      security: [{cookieAuth: []}, {bearerAuth: []}]
      responses:
        '200':
          description: Capability map based on the current database role
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminCapability'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /admin/users:
    get:
      tags: [Admin]
      summary: Search and paginate users without credentials or private content
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - name: search
          in: query
          required: false
          schema: {type: string, maxLength: 128}
        - name: status
          in: query
          required: false
          schema: {type: string, enum: [ACTIVE, DISABLED]}
        - name: role
          in: query
          required: false
          schema: {type: string, enum: [STUDENT, TEACHER, ADMIN]}
      responses:
        '200':
          description: User page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminUserPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /admin/users/{id}:
    get:
      tags: [Admin]
      summary: Return a safe administrator view of one user
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/AdminUserId'}
      responses:
        '200':
          description: User detail without password hashes, verification codes, tokens, or private resources
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminUser'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /admin/users/{id}/status:
    patch:
      tags: [Admin]
      summary: Enable or disable a user and create an audit event
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/AdminUserId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminUserStatusRequest'}
      responses:
        '200':
          description: Updated user state
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminUser'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /admin/users/{id}/roles:
    patch:
      tags: [Admin]
      summary: Replace a user's role set under server-side safeguards and audit it
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/AdminUserId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminUserRolesRequest'}
      responses:
        '200':
          description: Updated role set
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminUser'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /admin/audit-events:
    get:
      tags: [Admin]
      summary: Query non-sensitive administrator audit events
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - name: size
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
        - name: actorUserId
          in: query
          required: false
          schema: {type: integer, format: int64, minimum: 1}
        - name: action
          in: query
          required: false
          schema: {type: string, maxLength: 64}
        - name: targetType
          in: query
          required: false
          schema: {type: string, maxLength: 64}
        - name: targetId
          in: query
          required: false
          schema: {type: string, maxLength: 160}
        - name: from
          in: query
          required: false
          schema: {type: string, format: date-time}
        - name: to
          in: query
          required: false
          schema: {type: string, format: date-time}
      responses:
        '200':
          description: Audit event page with request IDs and state summaries only
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminAuditEventPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /admin/background-tasks:
    get:
      tags: [Admin]
      summary: List durable background task records
      description: >-
        The current writable task surface supports only `STALE_TASK_RECOVERY`.
        The ledger may still expose pre-existing records with another well-formed
        type, but retry and cancel reject them.
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - name: status
          in: query
          required: false
          schema: {type: string, enum: [PENDING, RUNNING, SUCCEEDED, FAILED, CANCELED]}
        - name: taskType
          in: query
          required: false
          description: Filter a stored task type. `STALE_TASK_RECOVERY` is the only currently submitted and executable task type.
          schema: {type: string, pattern: '^[A-Za-z0-9_.-]{1,64}$'}
      responses:
        '200':
          description: Task ledger page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BackgroundTaskPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /admin/background-tasks/{id}:
    get:
      tags: [Admin]
      summary: Return one durable background task record
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/BackgroundTaskId'}
      responses:
        '200':
          description: Task lifecycle, deadline, failure, retry, and request metadata
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BackgroundTask'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /admin/background-tasks/recover-timeouts:
    post:
      tags: [Admin]
      summary: Submit the only supported timeout-recovery task and return its durable record
      description: Creates a `STALE_TASK_RECOVERY` task. It marks expired pending or running work as failed; terminal success reports the affected count in `resultCount`.
      security: [{cookieAuth: []}, {bearerAuth: []}]
      responses:
        '202':
          description: Recovery submitted for asynchronous execution; terminal success records the recovered task count
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BackgroundTask'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /admin/background-tasks/{id}/retry:
    post:
      tags: [Admin]
      summary: Retry an eligible failed stale-task-recovery record
      description: Only a failed `STALE_TASK_RECOVERY` record below its attempt limit can be retried. Other task types return HTTP 409.
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/BackgroundTaskId'}
      responses:
        '202':
          description: Retry submitted for asynchronous execution
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BackgroundTask'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /admin/background-tasks/{id}/cancel:
    post:
      tags: [Admin]
      summary: Cancel a pending stale-task-recovery record
      description: Only a pending `STALE_TASK_RECOVERY` record can be canceled. Other task types return HTTP 409.
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/BackgroundTaskId'}
      responses:
        '200':
          description: Canceled task record
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BackgroundTask'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ds_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: {type: integer, minimum: 0, default: 0}
    PageSize:
      name: size
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    AdminUserId:
      name: id
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
    BackgroundTaskId:
      name: id
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
  responses:
    BadRequest:
      description: Invalid request
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Unauthorized:
      description: Authentication required or credentials invalid
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Forbidden:
      description: Authenticated user cannot access the requested source
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    NotFound:
      description: Resource not found or not visible to this user
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Conflict:
      description: State conflict
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
...
    TooManyRequests:
      description: Rate limit exceeded
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
...
  schemas:
    ApiError:
      type: object
      required: [code, message, requestId, details]
      properties:
        code: {type: string}
        message: {type: string}
        requestId: {type: string}
        details:
          type: array
          items: {type: string}
    RequestCodeRequest:
      type: object
      required: [email, purpose]
      properties:
        email: {type: string, format: email}
        purpose: {type: string, enum: [register, reset]}
    VerificationCodeDelivery:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Generic response. The verification code is delivered out of band and is never returned by HTTP.
    RegisterRequest:
      type: object
      required: [email, code, password]
      properties:
        email: {type: string, format: email}
        code: {type: string, minLength: 6, maxLength: 6}
        password: {type: string, minLength: 8, maxLength: 128}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, maxLength: 128}
    ResetPasswordRequest:
      type: object
      required: [email, code, password]
      properties:
        email: {type: string, format: email}
        code: {type: string, minLength: 6, maxLength: 6}
        password: {type: string, minLength: 8, maxLength: 128}
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: {type: string}
        user: {$ref: '#/components/schemas/User'}
    User:
      type: object
      required: [id, email, roles]
      properties:
        id: {type: integer, format: int64}
        email: {type: string, format: email}
        roles:
          type: array
          items: {type: string, enum: [STUDENT, TEACHER, ADMIN]}
...
    AdminCapability:
      type: object
      required: [userId, roles, modules, service]
      properties:
        userId: {type: integer, format: int64}
        roles:
          type: array
          items: {type: string, enum: [STUDENT, TEACHER, ADMIN]}
        modules:
          type: object
          additionalProperties: {$ref: '#/components/schemas/AdminModuleCapability'}
          description: Includes users, audit, reviewQueue, backgroundTasks, and modelSettings.
        service: {$ref: '#/components/schemas/AdminServiceStatus'}
    AdminModuleCapability:
      type: object
      required: [available, status]
      properties:
        available: {type: boolean}
        status: {type: string, enum: [AVAILABLE, UNAVAILABLE, NOT_CONFIGURED]}
        reason: {type: string, nullable: true}
    AdminServiceStatus:
      type: object
      required: [name, version, status]
      properties:
        name: {type: string, enum: [spring]}
        version: {type: string}
        status: {type: string, enum: [AVAILABLE, UNAVAILABLE]}
    AdminUser:
      type: object
      required: [id, email, status, roles, createdAt, updatedAt]
      properties:
        id: {type: integer, format: int64}
        email: {type: string, format: email}
        status: {type: string, enum: [ACTIVE, DISABLED]}
        disabledReason: {type: string, nullable: true}
        disabledAt: {type: string, format: date-time, nullable: true}
        roles:
          type: array
          items: {type: string, enum: [STUDENT, TEACHER, ADMIN]}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    AdminUserPage:
      type: object
      required: [items, page, size, total]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/AdminUser'}
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, format: int64, minimum: 0}
    AdminUserStatusRequest:
      type: object
      required: [status]
      properties:
        status: {type: string, enum: [ACTIVE, DISABLED]}
        reason: {type: string, maxLength: 500, nullable: true}
    AdminUserRolesRequest:
      type: object
      required: [roles]
      properties:
        roles:
          type: array
          minItems: 1
          maxItems: 3
          uniqueItems: true
          items: {type: string, enum: [STUDENT, TEACHER, ADMIN]}
    AdminAuditEvent:
      type: object
      required: [id, actorUserId, action, targetType, targetId, result, requestId, beforeSummary, afterSummary, createdAt]
      properties:
        id: {type: integer, format: int64}
        actorUserId: {type: integer, format: int64}
        action: {type: string}
        targetType: {type: string}
        targetId: {type: string}
        result: {type: string}
        requestId: {type: string, maxLength: 128}
        beforeSummary: {type: string, description: Safe state summary; never credentials or private body content.}
        afterSummary: {type: string, description: Safe state summary; never credentials or private body content. Model-config connection-test entries include non-negative connectionTestElapsedMs and credentialRedacted=true.}
        createdAt: {type: string, format: date-time}
    AdminAuditEventPage:
      type: object
      required: [items, page, size, total]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/AdminAuditEvent'}
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, format: int64, minimum: 0}
    BackgroundTask:
      type: object
      required: [id, taskType, status, createdAt, retryCount, maxAttempts, requestId]
      properties:
        id: {type: integer, format: int64}
        taskType:
          type: string
          maxLength: 64
          description: >-
            STALE_TASK_RECOVERY is the only current submitted, executable,
            retryable, and cancelable type. The read-only ledger can contain
            older records with another type.
        status: {type: string, enum: [PENDING, RUNNING, SUCCEEDED, FAILED, CANCELED]}
        createdAt: {type: string, format: date-time}
        startedAt: {type: string, format: date-time, nullable: true}
        deadlineAt: {type: string, format: date-time, nullable: true}
        heartbeatAt: {type: string, format: date-time, nullable: true}
        finishedAt: {type: string, format: date-time, nullable: true}
        failureCode: {type: string, nullable: true}
        failureReason: {type: string, nullable: true}
        resultCount: {type: integer, minimum: 0, nullable: true, description: Number of records affected by a completed maintenance task}
        retryCount: {type: integer, minimum: 0}
        maxAttempts: {type: integer, minimum: 1}
        cancelRequestedAt: {type: string, format: date-time, nullable: true}
        requestedByUserId: {type: integer, format: int64, nullable: true}
        requestId: {type: string, maxLength: 128}
    BackgroundTaskPage:
      type: object
      required: [items, page, size, total]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/BackgroundTask'}
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, format: int64, minimum: 0}
...